Start work on updates and having dawn/runtime/hibike run as services.
Everything is pretty rough right now but still should be in a working
state. Note that these changes are currently untested. Testing on the VM
will be done shortly, and testing on the Beaglebone blacks will be done
more or less immediately after the break.

Also remembered to add installation of node modules in
master_frank_setup.sh