Remove dawn-related stuff.
Since dawn is most likely going to be running as a desktop app that's
downloaded / installed instead of being served from the robot, there's
no need for quite a bit of config that was originally here. Got started
on removing the extraneous dawn-related stuff in case this change goes
through.