Add an update deployment script.
The deployment script packages a new update, signs the tarball, uploads
it to the robot, and then uses the robot's update script (same as the
one used for updates deployed through dawn) to install it.

Note that the private keys for signing the tarball as well as ssh-ing
into the robot are required to properly run this script.