Ensure updates are (actually) written to disk.
Thanks to Ryan we pretty quickly debugged the mysterious problem that
was bricking the beaglebones of some unfortunate teams (at least I hope).

The current theory about what was going on is that the new private key
was copied to the robot but not written back to disk (merp ext4), and
since the robots are often reset by cutting the power completely this
left the robot in a state with no ssh keys (and sometimes without
runtime).